move ChannelDecision to a common type
